Refactor admin_controller
